+++ yandex/resource_yandex_compute_instance_group.go
@@ -398,7 +398,6 @@ func resourceYandexComputeInstanceGroup() *schema.Resource {
 							Type:     schema.TypeList,
 							MaxItems: 1,
 							Optional: true,
-							Computed: true,
 							Elem: &schema.Resource{
 								Schema: map[string]*schema.Schema{
 									"placement_group_id": {