+++ CHANGELOG.md
@@ -1,14 +1,19 @@
-## 0.56.0 (Unreleased)
+## 0.57.0 (Unreleased)
 FEATURES:
+* support k8s node group placement groups both in resource and data source.
 * **New Data Source:** `yandex_alb_target_group`
 * **New Resource:** `yandex_alb_target_group`
 
+## 0.56.0 (April 15, 2021)
 ENHANCEMENTS:
 * dataproc: supported `security_group_ids`
 * add `dns_record`, `ipv6_dns_record` and `nat_dns_record` to `network_interface` entity in `yandex_compute_instance_group`
 * ydb: support for Yandex Database clusters
 * compute: increase disk size limit from 4096Gb to 8192Gb 
 * vpc: add `name` field description at vpc_security_group datasource and example of it usage
+
+BUG FIXES:
+* compute: placement_policy update in `yandex_compute_instance_group`
 
 ## 0.55.0 (April 1, 2021)
 FEATURES: